refactored router to allow for more components
