+++ node-backend-starter-code/src/js/components/App.jsx
@@ -1,14 +1,16 @@
 import React, {Component} from 'react'
-import SearchMovies from '../containers/SearchMovies.jsx'
-import VisibleMoviesList from '../containers/VisibleMoviesList.jsx'
+import { Link } from 'react-router'
 
 export default class App extends Component {
   render () {
     return (
       <div>
-        <SearchMovies />
-        <VisibleMoviesList />
+        <nav>
+          <Link to='/'>Search &nbsp;</Link>
+          <Link to='/favorites'>Favorites</Link>
+        </nav>
+        {this.props.children}
       </div>
-      )
+    )
   }
 }
