+++ node-backend-starter-code/src/js/components/MovieDetail.jsx
@@ -1,0 +1,10 @@
+import React from 'react'
+import { Link } from 'react-router'
+
+const MoviesDetail = ({movie}) => {
+  return (
+    <div> Movie here! </div>
+  )
+}
+
+export default MoviesDetail
