+++ node-backend-starter-code/src/js/main.jsx
@@ -2,33 +2,21 @@ import '../scss/styles.scss'
 
 import React from 'react'
 import { render } from 'react-dom'
-import { createStore, applyMiddleware, compose } from 'redux'
 import { Provider } from 'react-redux'
-import { Router, Route } from 'react-router'
+import { Router, Route, IndexRoute } from 'react-router'
 import createBrowserHistory from 'history/lib/createBrowserHistory'
-import reducers from './reducers'
-import thunkMiddleware from 'redux-thunk'
-import createLogger from 'redux-logger'
+import store from './store'
 import App from './components/App.jsx'
-
-let initialState = {
-  selectedTab: 'search',
-  // favorites: [],
-  movies: []
-}
-
-const store = createStore(
-  reducers,
-  initialState,
-  compose(applyMiddleware(thunkMiddleware, createLogger()),
-    window.devToolsExtension ? window.devToolsExtension() : f => f
-  )
-)
+import Search from './components/Search.jsx'
+import MovieDetail from './components/MovieDetail.jsx'
 
 render(
   <Provider store={store}>
     <Router history={createBrowserHistory()}>
-      <Route path='/' component={App} />
+      <Route path='/' component={App}>
+        <IndexRoute component={Search}/>
+        <Route path='movie/:id' component={MovieDetail} />
+      </Route>
     </Router>
   </Provider>,
   document.getElementById('app')
